SEM5/AWS/ass5/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
